Add packet-drop to fix the new flaky failover test (#2196)
The new test was added in #2178, obviously there may be
pending reads in the connection, so there may be a race
in the DROP-CLUSTER-PACKET-FILTER part causing the test
to fail. Add CLOSE-CLUSTER-LINK-ON-PACKET-DROP to ensure
that the replica does not process the packet.

Signed-off-by: Binbin <binloveplay1314@qq.com>
(cherry picked from commit 2019337e7442a1421955716bb7663801a2f29c8d)
Signed-off-by: Viktor Söderqvist <viktor.soderqvist@est.tech>